requirements.txt
appdirs==1.4.4
asgiref==3.8.1
baron==0.10.1
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==5.0.3
django-cache-url==3.4.5
django-environ==0.11.2
djangorestframework==3.15.1
environs==11.0.0
fire==0.6.0
fonttools==4.50.0
lxml==5.2.0
marshmallow==3.21.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
packaging==24.0
pdf2docx==0.5.8
PyMuPDF==1.24.0
PyMuPDFb==1.24.0
python-docx==1.1.0
python-dotenv==1.0.1
redbaron==0.9.2
rply==0.7.8
six==1.16.0
sqlparse==0.4.4
termcolor==2.4.0
typing_extensions==4.10.0
